+++ src/pluginx/trader_ape/set_field_ape.h
@@ -32,18 +32,18 @@ public:
 	~SetField();
 
 public:
-	typedef bool (SetField::*SetFieldFunc)( int64_t api_session, Request* request );
+	typedef bool (SetField::*SetFieldFunc)( int32_t api_session, Request* request );
 
-	bool SetField_120001_620001( int64_t api_session, Request* request ); // 单个委托下单
-	bool SetField_120002_620021( int64_t api_session, Request* request ); // 单个委托撤单
-	bool SetField_120003_620002( int64_t api_session, Request* request ); // 批量委托下单
-	bool SetField_120004_620022( int64_t api_session, Request* request ); // 批量委托撤单
-	bool SetField_130002_630002( int64_t api_session, Request* request ); // 查询客户资金
-	bool SetField_130004_630004( int64_t api_session, Request* request ); // 查询客户持仓
-	bool SetField_130005_630005( int64_t api_session, Request* request ); // 查询客户当日委托
-	bool SetField_130006_630006( int64_t api_session, Request* request ); // 查询客户当日成交
-	bool SetField_130008_601410( int64_t api_session, Request* request ); // 查询ETF基本信息
-	bool SetField_130009_601411( int64_t api_session, Request* request ); // 查询ETF成分股信息
+	bool SetField_120001_620001( int32_t api_session, Request* request ); // 单个委托下单
+	bool SetField_120002_620021( int32_t api_session, Request* request ); // 单个委托撤单
+	bool SetField_120003_620002( int32_t api_session, Request* request ); // 批量委托下单
+	bool SetField_120004_620022( int32_t api_session, Request* request ); // 批量委托撤单
+	bool SetField_130002_630002( int32_t api_session, Request* request ); // 查询客户资金
+	bool SetField_130004_630004( int32_t api_session, Request* request ); // 查询客户持仓
+	bool SetField_130005_630005( int32_t api_session, Request* request ); // 查询客户当日委托
+	bool SetField_130006_630006( int32_t api_session, Request* request ); // 查询客户当日成交
+	bool SetField_130008_601410( int32_t api_session, Request* request ); // 查询ETF基本信息
+	bool SetField_130009_601411( int32_t api_session, Request* request ); // 查询ETF成分股信息
 
 public:
 	basicx::SysLog_S* m_syslog;
